refactor(extension): 💡 slash
