+++ packages/extensions/slash/src/index.ts
@@ -1,6 +1,5 @@
 import Slash from './slash'
 
 export * from './slash'
-export * from './slash-zh'
 
 export default Slash
