+++ packages/extensions/unique-id/src/index.ts
@@ -1,5 +1,5 @@
-import { UniqueID } from './unique_id'
+import { UniqueID } from './unique-id'
 
-export * from './unique_id'
+export * from './unique-id'
 
 export default UniqueID
