+++ packages/vue/src/extensions/index.ts
@@ -1,6 +1,7 @@
 export { default as OColorHighlighter } from '@yiitap/extension-color-highlighter'
 export { default as OFocus } from '@yiitap/extension-focus'
 export { default as OPlaceholder } from '@yiitap/extension-placeholder'
+export { default as OSlash } from '@yiitap/extension-slash'
 export { default as OTable } from '@yiitap/extension-table'
 export { default as OTrailingNode } from '@yiitap/extension-trailing-node'
 export { default as OUniqueID } from '@yiitap/extension-unique-id'
